Move GitHub API calls to a separate class
